SaloniNavgire_114_B.TechA_Digimark/requirements.txt
Flask
Werkzeug
Pillow
pytesseract
opencv-python
numpy
pandas
matplotlib
nltk
scikit-learn
pdf2image
fpdf
flask-wtf